Move back testcomponents to pkg/flow/internal
