+++ pkg/flow/internal/testcomponents/singleton.go
